+++ roslibrust/examples/generic_client_services.rs
@@ -41,7 +41,10 @@ async fn main() {
         async fn run(self) {
             let _handle = self
                 .ros
-                .advertise_service::<std_srvs::SetBool, _>("/my_set_bool", Self::handle_service)
+                .advertise_service::<std_srvs::SetBool, _>(
+                    "/my_set_bool",
+                    Self::handle_service,
+                )
                 .await
                 .unwrap();