+++ roslibrust_ros2/examples/service_server.rs
@@ -33,7 +33,10 @@ async fn main() {
 
     // Server will be running for as long as _server is kept alive
     let _server = client
-        .advertise_service::<roslibrust_test::ros2::std_srvs::SetBool, _>("/set_bool", server_fn)
+        .advertise_service::<roslibrust_test::ros2::std_srvs::SetBool, _>(
+            "/set_bool",
+            server_fn,
+        )
         .await
         .unwrap();
     // Server is now up and can be called via command line: